fix: move ogImage to components folder
